Don’t try to delete articles that don’t exist

+++ src/Listeners/DeleteArticle.php
@@ -2,13 +2,29 @@
 
 namespace Aerni\AppleNews\Listeners;
 
-use Aerni\AppleNews\Facades\Article;
 use Statamic\Events\EntryDeleted;
+use Aerni\AppleNews\Facades\Article;
+use Statamic\Contracts\Entries\Entry;
 
 class DeleteArticle
 {
     public function handle(EntryDeleted $event): void
     {
-        Article::delete($event->entry);
+        if ($this->shouldDeleteArticle($event->entry)) {
+            Article::delete($event->entry);
+        }
+    }
+
+    protected function shouldDeleteArticle(Entry $entry): bool
+    {
+        if ($entry->get('apple_news_published')) {
+            return false;
+        }
+
+        if (! $entry->get('apple_news_id')) {
+            return false;
+        }
+
+        return true;
     }
 }
